modify service

+++ client/src/store/entities/service.ts
@@ -2,6 +2,8 @@ import { Dispatch } from "react";
 import { sliceMap } from '.'
 import API from 'api'
 
+export const confirmDelete = () => window.confirm('Вы действительно хотите продолжить удаление? Отменить операцию будет нельзя!')
+
 export const getActions = (name: EntityName) => sliceMap[name].actions
 
 export const readAllEntities = () => async (dispatch: Dispatch<any>) => {
@@ -53,7 +55,7 @@ export const updateEntity = (name: EntityName) => async (dispatch: Dispatch<any>
 }
 
 export const deleteEntity = (name: EntityName, id: number) => async (dispatch: Dispatch<any>, getState: GlobalStateGetter) => {
-    if (window.confirm('Вы действительно хотите продолжить удаление? Отменить операцию будет нельзя!')) {
+    if (confirmDelete()) {
         const { removeEntity } = getActions(name)
         await API[name].delete(id)
         dispatch(removeEntity(id))
@@ -65,7 +67,7 @@ export const attachLabel = (cardId: number, labelId: number) => async (dispatch:
     const { entities } = getState().labels
     const relatedLabel = entities.find(e => e.id === labelId)
     const response = await API.cards.addLabel(cardId, labelId)
-    console.log(response)
+    console.log('[DEBUG] [attach-label]', response)
     dispatch(addLinkedEntity({
         parentId: cardId,
         payload: relatedLabel,
@@ -76,7 +78,7 @@ export const attachLabel = (cardId: number, labelId: number) => async (dispatch:
 export const detachLabel = (cardId: number, labelId: number) => async (dispatch: Dispatch<any>) => {
     const { removeLinkedEntity } = getActions('cards')
     const response = await API.cards.deleteLabel(cardId, labelId)
-    console.log(response)
+    console.log('[DEBUG] [detach-label]', response)
     dispatch(removeLinkedEntity({
         parentId: cardId,
         childId: labelId,
@@ -90,12 +92,12 @@ export const addComment = (cardId: number) => async (dispatch: Dispatch<any>, ge
             resolve(response.data)
         }))
     })
-    console.log('cm:', responseEntity)
+    console.log('[DEBUG] [add-comment] create', responseEntity)
     const { addLinkedEntity } = getActions('cards')
     const { entities } = getState().comments
     const relatedComment = entities.find(e => e.id === responseEntity.id)!
     const response = await API.cards.addComment(cardId, relatedComment.id)
-    console.log(response)
+    console.log('[DEBUG] [add-comment] attach', response)
     dispatch(addLinkedEntity({
         parentId: cardId,
         payload: relatedComment,
@@ -104,14 +106,16 @@ export const addComment = (cardId: number) => async (dispatch: Dispatch<any>, ge
 }
 
 export const deleteComment = (cardId: number, commentId: number) => async (dispatch: Dispatch<any>) => {
-    const { removeLinkedEntity } = getActions('cards')
-    const responseDetach = await API.cards.deleteComment(cardId, commentId)
-    console.log(responseDetach)
-    const responseDelete = await API.comments.delete(commentId)
-    console.log(responseDelete)
-    dispatch(removeLinkedEntity({
-        parentId: cardId,
-        childId: commentId,
-        childName: 'comments',
-    }))
+    if (confirmDelete()) {
+        const { removeLinkedEntity } = getActions('cards')
+        const responseDetach = await API.cards.deleteComment(cardId, commentId)
+        console.log('[DEBUG] [delete-comment] detach', responseDetach)
+        const responseDelete = await API.comments.delete(commentId)
+        console.log('[DEBUG] [delete-comment] delete', responseDelete)
+        dispatch(removeLinkedEntity({
+            parentId: cardId,
+            childId: commentId,
+            childName: 'comments',
+        }))
+    }
 }
